+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Attachment as field type
+- Processing of an attachment as field type
 
 ## [1.3.0] - 2022-07-19